# app/web/openapi.yaml
openapi: 3.0.3
info:
  title: BirdLense Hub UI API
  version: 0.5.0
  description: |
    BirdLense Hub is an automated bird feeder monitoring and analytics system. It uses video and audio processing to detect, identify, and log bird (and other animal) visits, providing rich statistics, visualizations, and configuration options via a REST API. The API supports:
      - Querying real-time and historical bird visit data
      - Retrieving and managing video, species, and food information
      - Accessing system health, metrics, and storage usage
      - Configuring system settings (e.g., detection parameters, notification preferences)

    The API is organized into several groups:
      - UI routes: Endpoints for retrieving and managing bird, video, food, and visit data, as well as system configuration.
      - System routes: Endpoints for system health, metrics, activity logs, and storage management.
      - Settings: Endpoints for reading and updating the system configuration.

    All endpoints return JSON. PATCH/POST endpoints may update system state. No authentication is required by default, but this can be enabled in deployment.

    When MCP is enabled, the same REST surface is also available to **authorized MCP clients** ([Model Context Protocol](https://modelcontextprotocol.io/)) at `/mcp` on the hub: automation, IDE integrations, or custom tooling can query visits, status, and configuration or perform allowed operations under `MCP_TOKEN`.

...
servers:
...
paths:
...
  /api/debug/motion-preview:
    get:
      description: MOG2 / static calibration preview with mask overlay (settings password).
      parameters:
        - in: query
          name: camera_id
          schema:
            type: string
        - in: query
          name: mode
          schema:
            type: string
            enum:
              - detection_mog2
              - trigger_mog2
              - static
        - in: query
          name: overrides
          schema:
            type: string
            description: JSON overrides for processor/triggers.opencv (unsaved form values).
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Frame unavailable
      summary: Motion / static calibration preview
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...